+++ lib/passageidentity/user_api.rb
@@ -42,6 +42,40 @@ module Passage
       end
     end
 
+    def get_by_identifier(user_identifier:)
+      identifier_exists?(user_identifier)
+
+      begin
+        @req_opts[:limit] = 1
+        @req_opts[:identifier] = user_identifier.downcase
+        response = @user_client.list_paginated_users(@app_id, @req_opts)
+        users = response.users
+
+        if users.length() == 0
+          raise PassageError.new(
+                  message: "Passage User with identifer \"#{user_identifier}\" does not exist",
+                  status_code: 404,
+                  body: "user_not_found"
+                )
+        end
+        return get(user_id: users.first().id)
+      rescue Faraday::Error => e
+        if e.is_a? Faraday::ResourceNotFound
+          raise PassageError.new(
+                  message: "Passage User with identifer \"#{user_identifier}\" does not exist",
+                  status_code: e.response[:status],
+                  body: e.response[:body]
+                )
+        else
+          raise PassageError.new(
+                  message: "failed to get Passage User.",
+                  status_code: e.response[:status],
+                  body: e.response[:body]
+                )
+        end
+      end
+    end
+
     def activate(user_id:)
       user_exists?(user_id)
 
@@ -207,6 +241,12 @@ module Passage
     def user_exists?(user_id)
       if user_id.to_s.empty?
         raise PassageError.new(message: "must supply a valid user_id")
+      end
+    end
+
+    def identifier_exists?(identifier)
+      if identifier.to_s.empty?
+        raise PassageError.new(message: "must supply a valid identifier")
       end
     end